Bug #50612: Fix migrationscript to work with deactivated users
Fix check for univentionOffice365Enabled, only access attribute when
available
Also: remove the check for the ad connection UCRv. It is necessary to simplify
fixing broken customer environments - the joinscript can simply be
executed again without unsetting
office365/adconnection/alias/defaultADconnection.
The check itself is a remnant from the time the migration script initialized
the first ad connection itself. That is now done in the joinscript.