Bug #52840: Added library functions for MS Teams
Documentation can be found int he `README.graph`. It includes examples
on how to use the `terminaltest.py`.

> Separated the get_http_proxies function

This is a generic function, which is not office365 specific or specific
to the Azure API. I would describe it as a (generic) api_helper function
and named a file after that.

> Separated the `_get_client_assertion` from Azure

The function was nealy generic enough to be used for either of both: The
Azure and the Graph API. Minor modifications were made to get the file
loading out of the certificate generation (the payload includes an url,
which is different between Azure and Graph)

> Refactored the call_api function

During the attempt to adopt the existing call_api function to work with
Graph it turned out, that the pagination mechanism cannot be used the
same way as with Azure. That is because the url format has been
simplified with Microsoft Graph and has now the full URL as `next` link,
whereas with Azure it was necessary to use some urlbuilder.

The http headers are different as well. These could however be adopted.
A rewrite made reading the part a bit more convinient.

There was also getattrib used to call a certain method within the requests
class. That was unnecessary, because requests.request is a function,
which was already parameter compatible with the args.

In the end I went through the sanity checks, because I had to understand
how these would influence the pagination in the former implementation
(it is flow control...). The final result is a refactored call_api
method, which has been proven to work from the Graph class with Azure
calls. It could be backported nearly as it is, if you like the rewrite.

> Task #5627: get_all_aliases_from_ucr nochmal anschauen

Only the test program `terminaltest.py` uses the function at the moment
and that only for convinience reasons, so that one does not have to
manually enter an alias. The function was therefore migrated to the
terminaltest.py program. The `graph_auth.py` is now empty, because
there was only unused code in it, These were remainings of the
certificate_helper task #5629. The file was deleted.

> Task #5623: Implements a `call_api` function for Graph

> Task #5624: Access token refresh and Caching done

Fixes task #5624: access token refresh and cache

And unified the generic calls further, so that we have them closer to
what msal would do.

Migrated call_api from graph to azure and minor tweaks

Testing buildpackage ci job

> Task #5600: Finished README.graph and added it

> Task #5771: remove refactored call_api

I kept only its source code comment, because that does not alter any
code logic.

Added new files to `univention-office365.install`