Prooved to work as a drop-in-replacement for azure_handler.py
The idea is, that the `azure_handler.py` survives first. It is used as
base class for the new `graph.py`, so that it offers the same
functionality. We can however expand the class without altering the base
class and we can also overwrite functions within, so that the new
implementations against the Graph-API can take over the work of
azure_handler.py step by step. Our tests will ensure, that there is no
incompatibility in the new implementation.