Refactored the call_api function
During the attempt to adopt the existing call_api function to work with
Graph it turned out, that the pagination mechanism cannot be used the
same way as with Azure. That is because the url format has been
simplified with Microsoft Graph and has now the full URL as `next` link,
whereas with Azure it was necessary to use some urlbuilder.

The http headers are different as well. These could however be adopted.
A rewrite made reading the part a bit more convinient.

There was also getattrib used to call a certain method within the requests
class. That was unnecessary, because requests.request is a function,
which was already parameter compatible with the args.

In the end I went through the sanity checks, because I had to understand
how these would influence the pagination in the former implementation
(it is flow control...). The final result is a refactored call_api
method, which has been proven to work from the Graph class with Azure
calls. It could be backported nearly as it is, if you like the rewrite.