Task #5627: get_all_aliases_from_ucr nochmal anschauen
Only the test program `terminaltest.py` uses the function at the moment
and that only for convinience reasons, so that one does not have to
manually enter an alias. The function was therefore migrated to the
terminaltest.py program. The `graph_auth.py` is now empty, because
there was only unused code in it, These were remainings of the
certificate_helper task #5629. The file was deleted.