Bug #52840: Fixed return value 'id' of create_group
It is impossible to run all tests at the moment, because at the time of
writing none of the azure users has an assignedplan to it. The 402 test
will for that reason fail with:

> Teams is disabled for the tenant. Ensure the tenant has valid Office365
> subscription.

and I commit this change, because it was possible to test it
individually and that worked.