Bug #48389: Add valid HTTP response code to create team call
If a group already is a team in Azure and a Graph Call to create a team
from the group is pending, Graph responds with a HTTP 409 statuscode.
This means the group already is a team. We should not retry the call in
this case. The response code 409 was added to the expected return
values.