Bug #55217 fix: Make Connector react again to User Office365 Activate/Deactivate and ADconnection changes
before typecasting the attributes from `set` to `list` in e3e8166,
the connector was notified of all attributes.
That made this bug possible in the first place.

But now we have to add the Ldap attributes to the listener attributes.
This ensures that the connector again reacts to Activating and
Deactivating a user for Office365 aswell as adding or removing Azure AD
connections of the user.